+++ packages/client/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Fixed
+
+- Throw an error if the user provides an invalid custom signed-extension.
+
 ## 1.20.7 - 2025-11-19
 
 ### Fixed